requirements.txt.txt
streamlit>=1.25.0
transformers>=4.40.0
nltk>=3.8.0
matplotlib>=3.7.0
seaborn>=0.12.2
fpdf>=1.7.2
python-docx>=0.8.11
PyPDF2>=3.0.1
pandas>=2.0.3